[WAYANG-796] Organize SQL platform examples
