--- bigquery-setup/data.yaml
@@ -1,72 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-projects:
-- id: test-project
-  datasets:
-  - id: sales
-    tables:
-    - id: orders
-      columns:
-      - name: order_id
-        type: INTEGER
-      - name: region
-        type: STRING
-      - name: product
-        type: STRING
-      - name: amount
-        type: FLOAT
-      data:
-      - order_id: 1
-        region: APAC
-        product: Widget A
-        amount: 1500.0
-      - order_id: 2
-        region: EMEA
-        product: Widget B
-        amount: 800.5
-      - order_id: 3
-        region: AMER
-        product: Widget A
-        amount: 2200.0
-      - order_id: 4
-        region: APAC
-        product: Widget C
-        amount: 350.75
-      - order_id: 5
-        region: EMEA
-        product: Widget A
-        amount: 1100.0
-      - order_id: 6
-        region: AMER
-        product: Widget B
-        amount: 950.25
-      - order_id: 7
-        region: APAC
-        product: Widget B
-        amount: 1750.0
-      - order_id: 8
-        region: EMEA
-        product: Widget C
-        amount: 420.0
-      - order_id: 9
-        region: AMER
-        product: Widget C
-        amount: 680.5
-      - order_id: 10
-        region: APAC
-        product: Widget A
-        amount: 3000.0
