--- bigquery-setup/docker-compose.yml
@@ -1,43 +1,0 @@
----
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Stack: BigQuery Emulator (goccy/bigquery-emulator)
-#
-# Single container — no metastore, no object storage needed.
-# Data is seeded from data.yaml on startup and lives in memory.
-#
-# Ports:
-#   HTTP (REST API): http://localhost:9050
-#   gRPC (Storage API): localhost:9060
-
-services:
-
-  bigquery:
-    image: ghcr.io/goccy/bigquery-emulator:0.6.6
-    platform: linux/amd64
-    container_name: bigquery-emulator
-    ports:
-      - "9050:9050"
-      - "9060:9060"
-    volumes:
-      - ./data.yaml:/data.yaml
-    command: --project=test-project --data-from-yaml=/data.yaml
-    healthcheck:
-      test: ["CMD-SHELL", "bash -c ': >/dev/tcp/localhost/9050'"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
