--- platforms-setup-guides/trino-setup/docker-compose.yml
@@ -1,142 +1,0 @@
----
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Stack: Trino + Hive Metastore + MinIO (S3 storage)
-#
-# This is the battle-tested Trino + Iceberg local setup.
-# Hive Metastore (HMS) stores Iceberg table metadata over Thrift on port 9083.
-# MinIO provides S3-compatible object storage for Parquet data files.
-# Trino's Iceberg connector uses HMS as catalog and writes Parquet to MinIO.
-#
-# Ports:
-#   Trino:    http://localhost:8080   (UI + JDBC)
-#   MinIO S3: http://localhost:9000
-#   MinIO UI: http://localhost:9001   (minioadmin / minioadmin)
-#   HMS:      localhost:9083          (Thrift, internal)
-#   Postgres: localhost:5432          (HMS backing store)
-
-services:
-
-  # PostgreSQL (Hive Metastore backing database)
-  postgres:
-    image: postgres:15-alpine
-    container_name: trino-postgres
-    environment:
-      POSTGRES_DB: metastore
-      POSTGRES_USER: hive
-      POSTGRES_PASSWORD: hive
-    ports:
-      - "5432:5432"
-    volumes:
-      - postgres-data:/var/lib/postgresql/data
-    healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U hive -d metastore"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-
-  # MinIO (S3-compatible object storage)
-  minio:
-    image: minio/minio:latest
-    container_name: trino-minio
-    environment:
-      MINIO_ROOT_USER: minioadmin
-      MINIO_ROOT_PASSWORD: minioadmin
-    ports:
-      - "9000:9000"
-      - "9001:9001"
-    command: server /data --console-address ":9001"
-    volumes:
-      - minio-data:/data
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://localhost:9000/minio/health/live"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-
-  # Create the warehouse bucket before HMS starts
-  minio-init:
-    image: minio/mc:latest
-    container_name: trino-minio-init
-    depends_on:
-      minio:
-        condition: service_healthy
-    entrypoint: >
-      /bin/sh -c "
-        mc alias set local http://minio:9000 minioadmin minioadmin;
-        mc mb local/warehouse --ignore-existing;
-        echo 'bucket warehouse ready';
-        exit 0;
-      "
-
-  # Hive Metastore
-  # naushadh/hive-metastore is a minimal, pre-configured HMS image
-  # that supports S3-compatible storage via env vars.
-  metastore:
-    image: naushadh/hive-metastore:latest
-    container_name: trino-metastore
-    depends_on:
-      postgres:
-        condition: service_healthy
-      minio:
-        condition: service_healthy
-      minio-init:
-        condition: service_completed_successfully
-    ports:
-      - "9083:9083"
-    environment:
-      DATABASE_HOST: postgres
-      DATABASE_DB: metastore
-      DATABASE_USER: hive
-      DATABASE_PASSWORD: hive
-      # S3 / MinIO
-      S3_ENDPOINT_URL: http://minio:9000
-      S3_BUCKET: warehouse
-      S3_PREFIX: /
-      AWS_ACCESS_KEY_ID: minioadmin
-      AWS_SECRET_ACCESS_KEY: minioadmin
-      REGION: us-east-1
-    # No nc/curl in this image; use bash's /dev/tcp built-in
-    healthcheck:
-      test: ["CMD", "/bin/bash", "-c", "exec 3<>/dev/tcp/localhost/9083 2>/dev/null && exit 0 || exit 1"]
-      interval: 15s
-      timeout: 10s
-      retries: 15
-
-  # Trino
-  trino:
-    image: trinodb/trino:435
-    container_name: trino
-    depends_on:
-      metastore:
-        condition: service_healthy
-      minio:
-        condition: service_healthy
-    ports:
-      - "8080:8080"
-    volumes:
-      - ./trino/catalog:/etc/trino/catalog
-      - ./trino/config.properties:/etc/trino/config.properties
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://localhost:8080/v1/info"]
-      interval: 15s
-      timeout: 10s
-      retries: 10
-
-volumes:
-  postgres-data:
-  minio-data:
