--- presto-setup/etc/catalog/memory.properties
@@ -1,26 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# PrestoDB in-memory connector.
-# Backs the `memory` catalog used by the Wayang Presto integration tests:
-# tables created here (CREATE TABLE memory.<schema>.<table>) live in worker
-# memory and are dropped at teardown.
-connector.name=memory
-# Cap on-heap connector data per node. The image's default heap is -Xmx1G with
-# -XX:+ExitOnOutOfMemoryError, so keep this well under the heap; 256MB is ample
-# for the small, transient test tables. Raise the heap (custom jvm.config) before
-# increasing this if you reuse the stack for larger inserts.
-memory.max-data-per-node=256MB
