--- trino-setup/trino/catalog/iceberg.properties
@@ -1,32 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Iceberg catalog backed by Hive Metastore (HMS) and MinIO (S3 storage).
-# HMS stores Iceberg table metadata; Trino writes Parquet files to MinIO via S3.
-connector.name=iceberg
-iceberg.catalog.type=hive_metastore
-hive.metastore.uri=thrift://metastore:9083
-
-# Native S3 filesystem handles both s3:// and s3a:// (which HMS uses internally).
-# This avoids the "No FileSystem for scheme s3" error when HMS assigns locations.
-fs.native-s3.enabled=true
-s3.endpoint=http://minio:9000
-s3.path-style-access=true
-s3.aws-access-key=minioadmin
-s3.aws-secret-key=minioadmin
-s3.region=us-east-1
-
-iceberg.file-format=PARQUET
