+++ wayang-applications/src/test/java/org/apache/wayang/applications/SqlPlatformDemoTest.java
@@ -1,0 +1,61 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.wayang.applications;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+class SqlPlatformDemoTest {
+
+    @TempDir
+    Path directory;
+
+    @Test
+    void bigQueryCostModeLoadsWayangConfiguration() throws Exception {
+        Path configuration = directory.resolve("bigquery.properties");
+        Files.writeString(configuration, "wayang.bigquery.demo.table = `project.dataset.orders`\n");
+
+        assertDoesNotThrow(() -> BigQueryDemo.main(
+                new String[]{configuration.toUri().toString(), "cost"}
+        ));
+    }
+
+    @Test
+    void demosRequireConfigurationUrls() {
+        assertThrows(IllegalArgumentException.class, () -> BigQueryDemo.main(new String[0]));
+        assertThrows(IllegalArgumentException.class, () -> PrestoDemo.main(new String[0]));
+        assertThrows(IllegalArgumentException.class, () -> TrinoDemo.main(new String[0]));
+    }
+
+    @Test
+    void bigQueryOperatorModesRequireJdbcConfiguration() throws Exception {
+        Path configuration = directory.resolve("bigquery-no-jdbc.properties");
+        Files.writeString(configuration, "wayang.bigquery.demo.table = `project.dataset.orders`\n");
+
+        assertThrows(IllegalArgumentException.class, () -> BigQueryDemo.main(
+                new String[]{configuration.toUri().toString(), "filter"}
+        ));
+    }
+}
