+++ wayang-platforms/wayang-presto/src/main/resources/wayang-presto-defaults.properties
@@ -18,7 +18,7 @@
 # Connection (override per deployment in wayang.properties)
 # wayang.presto.jdbc.url      = jdbc:presto://localhost:8080
 # wayang.presto.jdbc.user     = test
-# wayang.presto.jdbc.password =
+# wayang.presto.jdbc.password = (optional; omit when no password is required)
 wayang.presto.jdbc.driverName = com.facebook.presto.jdbc.PrestoDriver
 
 # Hardware profile used by LoadProfileToTimeConverter.
